Rename `CodeGeneration` to `Annotation`.
