+++ gradle-plugins/model-compiler/src/main/java/io/spine/gradle/compiler/rejection/RejectionWriter.java
@@ -39,7 +39,7 @@ import java.nio.file.Files;
 import java.util.Map;
 
 import static com.squareup.javapoet.MethodSpec.constructorBuilder;
-import static io.spine.tools.CodeGeneration.generatedAnnotation;
+import static io.spine.tools.Annotation.generatedAnnotation;
 import static javax.lang.model.element.Modifier.FINAL;
 import static javax.lang.model.element.Modifier.PRIVATE;
 import static javax.lang.model.element.Modifier.PUBLIC;
