+++ gradle-plugins/model-compiler/src/main/java/io/spine/gradle/compiler/validate/ValidatingBuilderWriter.java
@@ -39,7 +39,7 @@ import java.nio.file.Files;
 
 import static io.spine.gradle.compiler.util.JavaSources.getBuilderClassName;
 import static io.spine.gradle.compiler.validate.ClassNames.getValidatorMessageClassName;
-import static io.spine.tools.CodeGeneration.generatedAnnotation;
+import static io.spine.tools.Annotation.generatedAnnotation;
 import static io.spine.util.Exceptions.newIllegalArgumentException;
 
 /**
